test: split skill pack workflow tests for lint
