+++ scripts/ci-check.sh
@@ -1,0 +1,31 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
+ROOT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")/.." && pwd)"
+cd "${ROOT_DIR}"
+
+npm ci --prefix .lint
+npm run --prefix .lint lint:lines
+go test ./...
+
+packages="$(
+	go list -f '{{if .TestGoFiles}}{{.ImportPath}}{{end}}' ./internal/... |
+		sed '/^$/d' |
+		grep -Ev '/repository$|/storage/(neo4j|postgres|redis)$'
+)"
+
+printf '%s\n' "${packages}"
+go test ${packages} -covermode=atomic -coverprofile=coverage.out
+go tool cover -func=coverage.out
+
+total="$(go tool cover -func=coverage.out | awk '/^total:/ { gsub(/%/, "", $3); print $3 }')"
+awk \
+	-v total="${total}" \
+	-v threshold="${COVERAGE_THRESHOLD:-90.0}" \
+	'BEGIN {
+		if ((total + 0) < (threshold + 0)) {
+			printf("coverage %.1f%% is below required %.1f%%\n", total, threshold)
+			exit 1
+		}
+		printf("coverage %.1f%% meets required %.1f%%\n", total, threshold)
+	}'
